app/requirements.txt
attrs==23.1.0
autocommand==2.2.2
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.1
exceptiongroup==1.1.2
Flask==3.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oathtool==2.3.1
outcome==1.2.0
packaging==23.1
path==16.10.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.19.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.10.0
urllib3==1.26.16
urllib3-secure-extra==0.1.0
uWSGI==2.0.24
webdriver-manager==4.0.1
Werkzeug==3.0.2
wsproto==1.2.0
